pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alumni-management-system"
version = "1.0.0"
description = "Alumni Management System with LinkedIn scraping, database storage, and web interface"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "IIM Ranchi Alumni Relations Committee"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "playwright>=1.40.0",
    "b2sdk>=1.24.0",
    "streamlit>=1.29.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
    "openai>=1.6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "flake8>=6.1.0",
    "black>=23.12.0",
    "mypy>=1.7.0",
]

[project.scripts]
alumni-frontend = "src.frontend.app:main"

[project.urls]
Homepage = "https://github.com/im45145v/ARC-IIMR"
Repository = "https://github.com/im45145v/ARC-IIMR"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
